chat PR pane: never auto-open; explicit open persists until closed
Removes the webhook-driven auto-pop of the floating PR pane on both the
ADE chat surface (AgentChatPane) and the CLI session surface
(WorkViewArea). The pane now opens only via an explicit toggle and,
because per-chat companion state persists as before, an opened pane
stays open across chat switches and restarts until closed.

- useChatPrAutoPop -> useChatPrPaneOpen: drops the prs.onEvent
  subscription and the getForLane baseline seed; keeps the hydrate/
  persist effects and their stale-flush guard intact.
- ChatPrPane: removes delta detection helpers, the transient delta
  banner, FieldPulse flashes, and their now-unused imports.
- Tests: persistence suite kept; pop/delta suites removed;
  WorkViewArea pin-routing assertions moved onto the pane props.
