--- apps/desktop/src/renderer/components/chat/useChatPrAutoPop.ts
@@ -1,122 +1,0 @@
-import { useEffect, useRef, useState, type Dispatch, type SetStateAction } from "react";
-
-import {
-  chatPrSignature,
-  detectChatPrDelta,
-  type ChatPrDelta,
-  type ChatPrSignature,
-} from "./ChatPrPane";
-import { patchChatCompanionUiState, readChatCompanionUiState } from "./chatCompanionUiState";
-import type { OpenProjectBinding } from "../../../shared/types";
-
-export type UseChatPrAutoPop = {
-  prPaneOpen: boolean;
-  setPrPaneOpen: Dispatch<SetStateAction<boolean>>;
-  /** The PR change that triggered the most recent auto-pop; handed to ChatPrPane. */
-  prPaneDelta: ChatPrDelta | null;
-};
-
-/**
- * Drives the floating PR pane's auto-pop for a lane's linked PR, shared by the
- * ADE chat surface (AgentChatPane) and the CLI session surface (WorkViewArea).
- *
- * On each webhook-driven `prs-updated` it re-reads this lane's PR summary and
- * pops the pane for a pop-worthy change — a newly created/linked PR, a lifecycle
- * change (merged/closed/reopened/ready/draft), or a new commit push. Checks and
- * review changes update the pane but do not pop it (see detectChatPrDelta). The
- * baseline is seeded silently so an already-open PR never pops on chat open.
- * Re-pops on each qualifying event; rapid bursts collapse into one visible pop
- * because the pane only ever shows the latest delta.
- *
- * `persistKey` (the surface's per-chat companion-state key) makes the pane's
- * open/closed state per chat AND durable across restarts: without it the pane
- * is bare component state, so every chat switch and every app launch reopens
- * from "closed" regardless of what the user left open.
- */
-export function useChatPrAutoPop(
-  laneId: string | null | undefined,
-  opts?: { persistKey?: string | null; runtimePin?: OpenProjectBinding | null },
-): UseChatPrAutoPop {
-  const persistKey = opts?.persistKey ?? null;
-  // The lane's PR lives on the lane's machine, so both the seed read and the
-  // event feed have to come from there — see `ChatGitToolbar.runtimePin`. Keyed
-  // on the pin KEY, read through a ref: a local pin is a fresh object on every
-  // cross-machine merge, and re-subscribing re-anchors the pinned event pump.
-  const runtimePin = opts?.runtimePin ?? null;
-  const runtimePinRef = useRef<OpenProjectBinding | null>(runtimePin);
-  runtimePinRef.current = runtimePin;
-  const runtimePinKey = runtimePin?.key ?? null;
-  const [prPaneOpen, setPrPaneOpen] = useState(
-    () => (persistKey ? readChatCompanionUiState(persistKey).prPaneOpen : false),
-  );
-  const [prPaneDelta, setPrPaneDelta] = useState<ChatPrDelta | null>(null);
-  const prevPrSigRef = useRef<ChatPrSignature | null>(null); // null = not yet seeded
-  const nonceRef = useRef(0);
-  // Which key the current `prPaneOpen` was hydrated from.
-  const hydratedPersistKeyRef = useRef<string | null>(persistKey);
-  // Set by the hydrate effect and consumed by the persist effect below. Both
-  // effects belong to the same fiber and run in declaration order within one
-  // passive-effect flush, so on the commit where `persistKey` changes the
-  // persist effect still closes over the OUTGOING chat's `prPaneOpen`. Marking
-  // the hydration here — rather than relying on the key ref, which the hydrate
-  // effect has already advanced by then — makes the persist effect skip exactly
-  // that one stale flush instead of writing chat A's value into chat B's record.
-  const pendingHydrationKeyRef = useRef<string | null>(null);
-
-  useEffect(() => {
-    if (hydratedPersistKeyRef.current === persistKey) return;
-    hydratedPersistKeyRef.current = persistKey;
-    pendingHydrationKeyRef.current = persistKey;
-    setPrPaneOpen(persistKey ? readChatCompanionUiState(persistKey).prPaneOpen : false);
-  }, [persistKey]);
-
-  // Every transition persists — the toolbar toggle, the ✕, and the webhook
-  // auto-pop below all land here because they all move the same state. The
-  // patch merges inside the store, so the chat shell's drawer fields on the
-  // same record survive.
-  useEffect(() => {
-    if (!persistKey || hydratedPersistKeyRef.current !== persistKey) return;
-    if (pendingHydrationKeyRef.current === persistKey) {
-      // Stale flush from the key change. If hydration changed the value, the
-      // corrective render re-runs this effect with the real one; if it didn't,
-      // storage already agrees and there is nothing to write either way.
-      pendingHydrationKeyRef.current = null;
-      return;
-    }
-    if (readChatCompanionUiState(persistKey).prPaneOpen === prPaneOpen) return;
-    patchChatCompanionUiState(persistKey, { prPaneOpen });
-  }, [persistKey, prPaneOpen]);
-
-  useEffect(() => {
-    setPrPaneDelta(null);
-    prevPrSigRef.current = null;
-    // The prs bridge can be absent in some surfaces / test harnesses / early
-    // boot; degrade to "no auto-pop" instead of crashing the host surface.
-    const prs = window.ade?.prs;
-    if (!laneId || !prs?.getForLane || !prs?.onEvent) return;
-    let cancelled = false;
-    prs
-      .getForLane(laneId, runtimePinRef.current)
-      .then((pr) => {
-        if (!cancelled && prevPrSigRef.current === null) prevPrSigRef.current = chatPrSignature(pr);
-      })
-      .catch(() => {});
-    const unsubscribe = prs.onEvent((event) => {
-      if (event.type !== "prs-updated") return;
-      const next = event.prs.find((pr) => pr.laneId === laneId) ?? null;
-      const prev = prevPrSigRef.current;
-      prevPrSigRef.current = chatPrSignature(next);
-      if (prev === null) return; // just seeded from this event — no pop
-      const change = detectChatPrDelta(prev, next);
-      if (!change) return;
-      setPrPaneDelta({ ...change, nonce: ++nonceRef.current });
-      setPrPaneOpen(true);
-    }, runtimePinRef.current);
-    return () => {
-      cancelled = true;
-      unsubscribe();
-    };
-  }, [laneId, runtimePinKey]);
-
-  return { prPaneOpen, setPrPaneOpen, prPaneDelta };
-}
